requirements.txt
arrow==1.2.2
asgiref==3.5.0
binaryornot==0.4.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cookiecutter==1.7.3
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.3
django-browser-reload==1.3.0
django-import-export==2.8.0
django-tailwind==3.1.1
et-xmlfile==1.1.0
idna==3.3
Jinja2==3.1.1
jinja2-time==0.2.0
MarkupPy==1.14
MarkupSafe==2.1.1
odfpy==1.4.1
openpyxl==3.0.9
poyo==0.5.0
psycopg2==2.9.3
python-dateutil==2.8.2
python-slugify==6.1.1
PyYAML==6.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tablib==3.2.0
text-unidecode==1.3
tzdata==2022.1
urllib3==1.26.9
xlrd==2.0.1
xlwt==1.3.0
